requirements.txt
Flask==2.2.3
gunicorn==23.0.0
gspread==6.2.1
google-auth==2.27.0
python-dateutil==2.8.2
